AI/requirements.txt
openai
pandas
pyodbc
python-dotenv